Fix bad assumptions in WebDAV interface
Specifically Jackrabbit's WebDAV assumes
the resource known whether it is a directory
or not even if it doesn't exist yet. This was
hard to work around. I ended up adding a fake
property in the input context to expose the
HTTP verb of the request. Thus I can respond
to MKCOL by creating a directory, while PUT will
create a file.